refactor(legacy): split ai/media/shortform bridge controllers
